feat: add transparency option
Enabled by pressing `T`, off at scene start.

+++ src/uv_mapping_bonus.c
@@ -6,7 +6,7 @@
 /*   By: fgargot <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/21 18:21:30 by fgargot           #+#    #+#             */
-/*   Updated: 2026/07/22 20:33:54 by fgargot          ###   ########.fr       */
+/*   Updated: 2026/07/24 23:19:27 by fgargot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -74,9 +74,9 @@ t_vec3	triangle_uv_to_color(t_object *obj, t_texture *tex, t_vec3 hit)
 	mlx_color	pixel;
 
 	uv = get_triangle_uv_hit(obj, hit);
-	uv.x = uv.x * tex->width - 1;
-	uv.y = (1 - uv.y) * tex->height - 1;
-	pixel = mlx_get_image_pixel(tex->mlx, tex->data, uv.x, uv.y);
+	uv.x = uv.x * (tex->width - 1);
+	uv.y = (1 - uv.y) * (tex->height - 1);
+	pixel = mlx_get_image_pixel(tex->mlx, tex->data, (int)uv.x, (int)uv.y);
 	return ((t_vec3){{pixel.r, pixel.g, pixel.b}});
 }
